==> Code/1_Clean_dataset.ipynb <==
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

## Read file and stock into DataFrame

df = pd.read_csv("../Data/chessDB.txt", sep = "###", skiprows = [0, 1, 2, 3, 4], header=None, 
                 engine='python')

print(df.shape)
print(df.size)
df.head()

##### We need to split the df, to keep the moves as 1 field and split the field 0 into several fields

moves_df = df[[1]]

games_df = df[[0]]

games_df = games_df[0].str.split(" ", expand=True)

games_df = games_df.drop(columns=0)

moves_df.columns = ['moves']

games_df['moves'] = moves_df['moves']

df = games_df

##### Let's name the columns to make it easier to work with the data

df = df.drop(columns = 16) #empty row

columns = ['date', 'result', 'white_elo', 'black_elo', '# moves', 'date_corrupt', 'result_corrupt',
           'Welo_corrupt', 'Belo_corrupt', 'eventDate_corrupt', 'setup_corrupt', 'fen_corrupt', 
           'result2_corrupt', 'out_of_year range', 'game_length_corrupt', 'moves']

df.columns = columns

df.head(2)

• date: date of the game <br> <br>
• result: 1-0 if white won, 0-1 if black won, 1/2-1/2 if draw and * if missing <br> <br>
• white_elo: the elo (strength) of the white player; the higher, the better the player <br> <br>
• black_elo: the elo (strength) of the black player; the higher, the better the player <br> <br>
• # moves: number of moves in the game <br> <br>
• date_corrupt: date_true if the date is corrupted in some way, date_false otherwise <br> <br>
• result_corrupt: result_true if the result of the game is somehow corrupted or missing, result_false otherwise <br> <br>
• Welo_corrupt: welo_true if the white elo is corrupted or missing, welo_false if not <br> <br>
• Belo_corrupt: belo_true if the black elo is corrupted or missing, belo_false if not <br> <br>
• evenDate_corrupt: edate_true if the date of the event where the game was held id corrupted (not recorded properly) or missing, edate_false otherwise <br> <br>
• setup_corrupt: setup_true if the game played was a Fischer Random Chess (a variant of chess where the pieces are setup randomly at the beginning of the game, not interesting here) or the initial setup is somehow corrupted, setup_false otherwise <br> <br>
• fen_corrupt: fen_true if a position recorder in the database is not legal in chess, making the game ior fen_false <br> <br>
• result2_corrupt: In the original file the result is provided in two places. At the end of each sequence of moves and in the attributes part. This flag indicates if the result is (is not) properly provided after the sequence of moves (just for checking consistency in the PGN file). <br> <br>
• out_of_year range: oyrange_false if the game was held in the range [1998, 2007].  <br> <br>
• game_length_corrupt: blen_true if the length of the game is not good, blen_false otherwise <br> <br>

## Remove irrelevant rows and columns

clean_df = df.copy(deep=True)

print(df[df.game_length_corrupt == 'blen_true']['moves'].unique())
clean_df = clean_df[clean_df.game_length_corrupt == 'blen_false']
# nothing but incorrect games, we remove them
# we also remove the row 'game_length_corrupt' as it is now irrelevant
clean_df = clean_df.drop(columns = 'game_length_corrupt')

# we don't care about the date of the game
clean_df = clean_df.drop(columns = ['date', 'date_corrupt', 'eventDate_corrupt', 'out_of_year range'])

len(clean_df[clean_df.fen_corrupt == 'fen_true'])

# the games where fen_corrupt is true do not correspond to possible games, the moves registered have errors
# thus we remove theses games (there are only 817 of them)
clean_df = clean_df[clean_df.fen_corrupt == 'fen_false']
# we may now remove the column 'fen_corrupt' as it is now always equal to fen_false
clean_df = clean_df.drop(columns = 'fen_corrupt')

# we check that when result_corrupt is true, result2_corrupt is also true
# thus we can only analyse one of the two
print(clean_df[clean_df.result_corrupt == 'result_true'].result2_corrupt.unique())

# we can also see that there are only 10 games where the result is corrupted
nb = len(clean_df[clean_df.result_corrupt == 'result_true'])
print(f"There are {nb} games where the result is corrupted")
# hence, we just get rid of them
clean_df = clean_df[clean_df.result_corrupt == 'result_false']
clean_df = clean_df.drop(columns = ['result_corrupt', 'result2_corrupt'])

# we don't really care whether the elo of either player is missing or not
clean_df = clean_df.drop(columns = ['Welo_corrupt', 'Belo_corrupt'])

clean_df.columns

clean_df['# moves'] = clean_df['# moves'].astype('int')

print(clean_df.setup_corrupt.unique())
#turns out that, after cleaning our dataset, there no row in clean_df where
#setup_corrupt is True, thus we can remove this field
clean_df = clean_df.drop(columns = 'setup_corrupt')

#we used to have a mixed type for the elos, with integers and string if missing
#we fix it, and put an negative elo (-1) if the elo happens to be missing
clean_df['white_elo'] = clean_df['white_elo'].replace('None', -1)
clean_df['black_elo'] = clean_df['black_elo'].replace('None', -1)

# let's also remove the games where the number of moves is quite low
# there aren't too many games like that, compared to the total amount of games in the database
nb = len(clean_df[clean_df["# moves"] <= 10])
print(f"There are {nb} games where the number of games is below 10")
clean_df = clean_df.loc[clean_df["# moves"] > 10]

clean_df

clean_df.to_csv("../Data/clean_df.csv", index=False)
