==> Code/2_Create_metric_multithread.ipynb <==
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
from tqdm import tqdm

import io
import pickle
import os
import glob

import chess.pgn
import chess.polyglot

df = pd.read_csv("../Data/clean_df.csv")

df

#number of moves in total throughout the dataset
nb_total_moves = df['# moves'].sum()
print(f"There are, in total, {nb_total_moves} moves in the database")

# the dataset is very big
# we will cut it to only take a small percentage of the whole dataset
# to do so, we'll only consider games where both elos are known and greater than 2500
# the idea is to only work with games played by the best players

len_prv_df = len(df)
df = df.loc[(df.white_elo > 2500) & (df.black_elo > 2500)]
print(f"There are now {len(df)} games in the dataset. Represents {round(100*len(df)/len_prv_df, 2)}% of the whole dataset.")

df

#number of moves we'll consider now
nb_total_moves = df['# moves'].sum()
print(f"We now have {nb_total_moves} moves in our dataset")

## Create a df where each line is a new move

#funtion to split a sequence of moves to a list of moves
def getMovesPGN(game_index):
    moves = df.iloc[game_index]['moves']
    if moves[0] == '':
        moves = moves[1:]
    pgn = io.StringIO(moves)
    return pgn

def getMovesSequence(game_index):
    pgn = getMovesPGN(game_index)
    game = chess.pgn.read_game(pgn)
    moves = []
    for move in game.mainline_moves():
        moves.append(move.uci())
    return moves    

#we create the game_index for the new database
#each game will be split in the amount of moves it has
game_index = []
moves = []
for i in tqdm(range(len(df))):
    nb_moves_game = df.iloc[i]['# moves']
    game_index += [i]*nb_moves_game
    moves_sequence = getMovesSequence(i)
    moves += moves_sequence

moves_df = pd.DataFrame()
moves_df['game_index'] = game_index
moves_df['moves'] = moves
moves_df['evaluation'] = [0]*len(game_index)
moves_df['fen'] = ['']*len(game_index)
moves_df['zobrist_key'] = ['']*len(game_index)
moves_df

moves_df.to_csv("../Data/moves_df.csv", index=False)

moves_df = pd.read_csv("../Data/moves_df.csv", dtype={"fen": str, 'zobrist_key': str})
moves_df.fillna('', inplace=True)

## Functions to create the metric

#### Write functions to evaluate a chess position

# functions to evaluate a given chess position, using a chess engine named stockfish

def stockfish_evaluation(board, time_limit = 0.01):
    #returns the score of the position, from the perspective of the white player
    #we will never change perspective throughout our work 
    engine = chess.engine.SimpleEngine.popen_uci("../Stockfish/stockfish_15_x64_avx2")
    result = engine.analyse(board, chess.engine.Limit(time=time_limit))
    engine.close()
    return result['score'].white()

def position_eval(board, time_limit = 0.01):
    # we need to consider scores where the engine has found a way to deliver mate differently
    # because is those cases, the score returned is a string, not an integer
    score = stockfish_evaluation(board, time_limit)
    if not score.is_mate():
        return score.score()
    #the engine has found a way to mate in a certain number of moves
    return mateScore(score)

def findNumberOfMovesBeforeMate(score):
    str_nb_moves_before_mate = ''
    s = str(score)
    i = len(s)-1
    while i >= 0 and s[i].isnumeric():
        str_nb_moves_before_mate = s[i] + str_nb_moves_before_mate
        i -= 1
    return int(str_nb_moves_before_mate)

def mateScore(score):
    nb_moves_before_mate = findNumberOfMovesBeforeMate(score)
    white_is_winning = (str(score.wdl())[9] == '1')
    # a.wdl() gives the probability of winning for white
    # if the 9th character is equal to 1, then white will deliver mate shortly
    # otherwise, white will lose soon and black is winning
    score_for_mate = 10000 - nb_moves_before_mate*100
    if not white_is_winning:
        score_for_mate = (-1)*score_for_mate
    return score_for_mate

# functions to play moves on a chess board

def getSquareNumber(square):
    #each square on the chess board is represented by an integer in the module chess that we use
    #this functions takes the name of a square as a parameter, for instance 'e2' and returns the integer associated
    col = square[0].lower()
    row = int(square[1])
    row_number = ord(col) - 97
    square_number = 8*(row-1) + row_number
    return square_number

def getMoveToPlay(startSquare, endSquare, promotion_piece=None):
    startSquare_num = getSquareNumber(startSquare)
    endStart_num = getSquareNumber(endSquare)
    move = chess.Move(startSquare_num, endStart_num, promotion = promotion_piece)
    return move

def getPieceToPromoteTo(move):
    #pieces : Pawn=1, Knight=2, Bishop=3, Rook=4, Queen=5, King=6
    if len(move) == 4:
        #a move is represented in our database with the square from which the piece starts 
        #and the square to which the piece goes, for instance 'e2e4'
        #if the move has a length of 4, there is no promotion happening 
        return None
    promotion_piece = move[-1]
    if promotion_piece == 'q':
        return 5
    if promotion_piece == 'r':
        return 4
    if promotion_piece == 'b':
        return 3
    if promotion_piece == 'n':
        return 2
    
def getMove(move):
    startSquare = move[:2]
    endSquare = move[2:4]
    if len(move) == 4:
        return getMoveToPlay(startSquare, endSquare)
    promotion_piece = getPieceToPromoteTo(move[-1])
    return getMoveToPlay(startSquare, endSquare, promotion_piece)

#function using the previous functions
#for a given game, the functions plays and evaluates each move of the game, 
#while also return a list of all the fen representations of each board during the game
#and the zobrist hashing of these boards
# a fen representation is a string that fully represents a chess position, with all the pieces, where they are
# and which player in currently playing. 
# a zobrist key is a way to represent a chess position with an 64-bits integer (theorized by the mathematician Zobrist)

def getGameEvalAndFen(game_index, dict_corres, nb_total_moves):
    board = chess.Board()
    index = dict_corres[game_index]
    
    evaluations = []
    fens = []
    zobrists = []
    while index < nb_total_moves and moves_df.iloc[index]['game_index'] == game_index:
        move = moves_df.iloc[index]['moves']
        move_to_play_on_board = getMove(move)
        board.push(move_to_play_on_board)
        eval_ = position_eval(board)
        evaluations.append(eval_)
        
        fen = board.fen()
        fens.append(fen)
        
        zobrist = str(chess.polyglot.zobrist_hash(board))
        zobrists.append(zobrist)
        
        index += 1
        
    return evaluations, fens, zobrists

#### Write functions to enable the use of multithreading

#to analyse the game which has the index game_index, we need to know what is the 
#index of the first move of this game in moves_df
#instead of searching everytime, we do it once and store the findings in a dictionary

def dictCorrespondanceIndex_gameIndex(moves_df):
    dict_corres = {}
    index = 0
    for index in tqdm(range(len(moves_df))):
        game_index = moves_df.iloc[index]['game_index']
        if dict_corres.get(game_index) == None:
            dict_corres[game_index] = index
    return dict_corres

# we first check whether the dictionary has already been calculated

try:
    with open('../Data/Metric_creation/dict_correspondance.pkl', 'rb') as f:
        dict_corres = pickle.load(f)
except:
    dict_corres = dictCorrespondanceIndex_gameIndex(moves_df)
    if not os.path.isdir('../Data/Metric_creation'):
        if not os.path.isdir('../Data'):
            os.mkdir('../Data')
        os.mkdir('../Data/Metric_creation')
    with open('../Data/Metric_creation/dict_correspondance.pkl', 'wb') as f:
        pickle.dump(dict_corres, f)

#calculating the metric for each game is a long process, and we may have to pause it and 
#resume later. Thus, we write this function to know what games have already been 
#analysed, and where to start to analyse new games

def gameIndexToStartAt(evals_dict, nb_total_games):
    if len(evals_dict) == 0:
        return 0
    games_already_analyzed = list(evals_dict.keys())
    if len(games_already_analyzed) == nb_total_games:
        return -1
    
    games_already_analyzed.sort()
    return games_already_analyzed[-1]+1    

nb_total_games = len(df)

# we will store the metric calculated in dictionaries
# we check whether the dictionaries already have been created or not

try:
    with open('../Data/Metric_creation/evaluations_dict.pkl', 'rb') as f:
        evals_dict = pickle.load(f)
except:
    evals_dict = {}
    
try:
    with open('../Data/Metric_creation/fens_dict.pkl', 'rb') as f:
        fens_dict = pickle.load(f)
except:
    fens_dict = {}
    
try:
    with open('../Data/Metric_creation/zobrists_dict.pkl', 'rb') as f:
        zob_dict = pickle.load(f)
except:
    zob_dict = {}

#as we use multithreading to speed the process, we want to avoid having different threads open the
#same dictionaries and writing into it, as it is not thread safe. 
#Thus, we create a new file for each game analysed, that we will pick up later on and add them 
#to the dictionaries. Don't worry, we will only create them 50 by 50 and erase them once read

def addToDict(game_index, nb_total_moves=nb_total_moves, dict_corres=dict_corres):
    
    evaluations, fens, zobrists = getGameEvalAndFen(game_index, dict_corres, nb_total_moves)
    with open('../Data/Metric_creation/evaluations_dict/' + str(game_index)+ '.pkl', 'wb') as f:
        pickle.dump(evaluations, f)
    with open('../Data/Metric_creation/fens_dict/' + str(game_index)+ '.pkl', 'wb') as f:
        pickle.dump(fens, f)
    with open('../Data/Metric_creation/zobrists_dict/' + str(game_index)+ '.pkl', 'wb') as f:
        pickle.dump(zobrists, f)

#we simply make sure that the paths we will use do exist

if not os.path.isdir('../Data/Metric_creation'):
    os.mkdir('../Data/Metric_creation')

if not os.path.isdir('../Data/Metric_creation/evaluations_dict'):
    os.mkdir('../Data/Metric_creation/evaluations_dict')
if not os.path.isdir('../Data/Metric_creation/fens_dict'):
    os.mkdir('../Data/Metric_creation/fens_dict')
if not os.path.isdir('../Data/Metric_creation/zobrists_dict'):
    os.mkdir('../Data/Metric_creation/zobrists_dict')

#this very lengthy function gets the files that have been created, containing the metric calculated
#and adds them to their respective dictionaries, then deletes them

def addFromFileToDict():
    eval_paths = glob.glob('../Data/Metric_creation/evaluations_dict/*.pkl')
    for eval_path in eval_paths:
        with open(eval_path, 'rb') as f:
            eval_ = pickle.load(f)
        index = int(eval_path.split(".")[-2].split("\\")[-1])
        evals_dict[index] = eval_
        os.remove(eval_path)
        
    fen_paths = glob.glob('../Data/Metric_creation/fens_dict/*.pkl')
    for fen_path in fen_paths:
        with open(fen_path, 'rb') as f:
            fen = pickle.load(f)
        index = int(fen_path.split(".")[-2].split("\\")[-1])
        fens_dict[index] = fen
        os.remove(fen_path)
        
    zob_paths = glob.glob('../Data/Metric_creation/zobrists_dict/*.pkl')
    for zob_path in zob_paths:
        with open(zob_path, 'rb') as f:
            zob = pickle.load(f)
        index = int(zob_path.split(".")[-2].split("\\")[-1])
        zob_dict[index] = zob
        os.remove(zob_path)
        
    with open('../Data/Metric_creation/evaluations_dict.pkl', "wb") as fp:   
        pickle.dump(evals_dict, fp)
    with open('../Data/Metric_creation/fens_dict.pkl', "wb") as fp:  
        pickle.dump(fens_dict, fp)
    with open('../Data/Metric_creation/zobrists_dict.pkl', "wb") as fp:   
        pickle.dump(zob_dict, fp)

## Analyse the games and create the metric

import concurrent.futures #allows the use of multithreading

threads_in_parallel = 50 # we'll have 50 threads working at the same time to improve efficiency

#we use a threadPoolExecutor to parallelize the executions of the metric creation per game
#makes it way faster to analyse each game

game_index = gameIndexToStartAt(evals_dict, nb_total_games)
print(f"Start at index: {game_index}")
with tqdm(total=nb_total_games-game_index) as pbar:
    while game_index < nb_total_games:
        with concurrent.futures.ThreadPoolExecutor(max_workers=threads_in_parallel) as executor:
            futures = []
            while game_index < nb_total_games and len(futures) < threads_in_parallel:
                futures.append(executor.submit(addToDict, game_index=game_index))
                game_index += 1
            
            for future in concurrent.futures.as_completed(futures):
                pbar.update(1)
            addFromFileToDict()

## Populate the database with our findings

## check that every game has been analyzed

for game_index in tqdm(range(nb_total_games)):
    if evals_dict.get(game_index) == None:
        addToDict(game_index)
addFromFileToDict()

# we populate our dataset moves_df with the dictionaries that we have calculated previously

def addToDataFrame(evals_dict, fens_dict, zob_dict):
    evals_ = []
    fens = []
    zobrists = []
    for i in range(len(evals_dict)):
        evals_i = evals_dict.get(i)
        fens_i = fens_dict.get(i)
        zobrists_i = zob_dict.get(i)
        
        evals_ += evals_i
        fens += fens_i
        zobrists += zobrists_i
    if len(evals_) < len(moves_df):
        evals_ += [0]*(len(moves_df)-len(evals_))
        fens += ['']*(len(moves_df)-len(fens))
        zobrists += ['']*(len(moves_df)-len(zobrists))
        
    moves_df['evaluation'] = evals_
    moves_df['fen'] = fens
    moves_df['zobrist_key'] = zobrists

addToDataFrame(evals_dict, fens_dict, zob_dict)

moves_df.to_csv("../Data/moves_df.csv", index=False)

moves_df

### Keep track of losses and wins

## We haven't kept track of who was the winner of each game, or if it is was a draw
#creating a new column in moves_df would not be efficient because for each move, we would have to keep track of the result 
#making it very redundant and this would make the dataset heavier for no reason
#hence, we create a dictionary that records the results, with the same idea as dict_corres

df

pd.unique(df.result)

def dict_game_results(df):
    df_wins = df.reset_index()
    dict_wins = {}
    for i in tqdm(range(len(df_wins))):
        result_i = df.iloc[i].result
        if result_i == '1-0':
            dict_wins[i] = 1
        elif result_i == '0-1':
            dict_wins[i] = 0
        else:
            dict_wins[i] = 0.5
            
    return dict_wins

# we first check whether the dictionary has already been calculated
try:
    with open('../Data/Metric_creation/dict_wins.pkl', 'rb') as f:
        dict_corres = pickle.load(f)
except:
    dict_wins = dict_game_results(df)
    if not os.path.isdir('../Data/Metric_creation'):
        if not os.path.isdir('../Data'):
            os.mkdir('../Data')
        os.mkdir('../Data/Metric_creation')
    with open('../Data/Metric_creation/dict_wins.pkl', 'wb') as f:
        pickle.dump(dict_wins, f)
